refactor: edit packages of NumberGenerator and input and output
