+++ src/main/java/racing/RacingMain.java
@@ -3,8 +3,8 @@ package racing;
 
 import racing.controller.RacingApplication;
 import racing.exception.RacingBaseException;
-import racing.generator.RandomNumberGenerator;
-import racing.input.RacingInput;
+import racing.model.generator.RandomNumberGenerator;
+import racing.view.input.RacingInput;
 
 public class RacingMain {
     public static void main(String[] args) {
