+++ src/main/java/racing/model/generator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racing.generator;
+package racing.model.generator;
 
 public interface NumberGenerator {
     int generate();
