+++ src/main/java/racing/model/generator/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racing.generator;
+package racing.model.generator;
 
 public class RandomNumberGenerator implements NumberGenerator {
