+++ src/test/java/racing/controller/RacingApplicationTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import racing.exception.IllegalCarNameException;
 import racing.exception.IllegalCountException;
-import racing.generator.SpecificNumberGenerator;
-import racing.input.RacingInput;
+import racing.model.generator.SpecificNumberGenerator;
+import racing.view.input.RacingInput;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
