+++ src/test/java/racing/view/input/RacingInputTest.java
@@ -1,7 +1,6 @@
-package racing.input;
+package racing.view.input;
 
 import org.junit.jupiter.api.Test;
-import racing.input.RacingInput;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
